Configure QNX build system for CLAPACK
This works on both QNX 7.1 and 8.0.

Compared to a normal project, there is one difference here: CLAPACK
requires `arithchk` to be executable on host in order to build. As
a result, we must build the binary for host before any `nto-` builds.

This is done by rewriting the main entry Makefile not to use recurse.mk
directly. Instead, we define our own Makefile targets for `nto-` and `linux-`,
and then specify the dependency -- all `nto-` targets must depend on
`linux-x86_64-o`. We manually recurse into the subdirectories within
build rules for those targets.

We then pass in the path to the compiled host binary when building the
`nto-*` targets.

QNXOSG-356
QNXOSG-357